refactor: move static_prompt module into traces/system_extraction
The queue producer/consumer and the extraction agent belong to the same
feature, so they now live in one module.
