+++ CLAUDE.md
@@ -337,8 +337,8 @@ Route groups: `(auth)/layout.tsx` requires a session (redirects to `/sign-in`); 
 
 ## Static System-Prompt Extraction (LAM-1899)
 
-- `app-server/src/static_prompt/` derives, per naive signature (`lmnr.span.prompt_hash`), a regex list whose matches are the prompt's dynamic parts. Distinct from `checkpoints/system_prompt.rs`, which resolves a single stripping regex inline on the checkpoints path — this pipeline is span-ingest-wide and cache-first.
-- Flow: the ingest producer (`traces/producer.rs::publish_span_messages`) collects `(prompt_hash, system_prompt)` from `preprocess_for_queue`, checks `static_prompt_regex:{project}:{hash}` and publishes misses to `static_prompt_queue` — best-effort, and only AFTER the span publish succeeded so rejected batches don't feed the accumulator. The consumer (`static_prompt/consumer.rs`) appends the prompt to `static_prompt_accumulator:{project}:{hash}` (24h TTL, capped at 10 samples so a persistently failing agent can't grow the cache value unboundedly); at ≥ 5 samples it takes the per-signature lock `static_prompt_lock:{project}:{hash}` (30-min TTL — runs are normally under 10 min, but the per-step upper bounds are high just in case), runs the extraction agent, writes the regex cache (7-day TTL), clears the accumulator, and releases the lock.
+- `app-server/src/traces/system_extraction/` (producer.rs + consumer.rs alongside the agent) derives, per naive signature (`lmnr.span.prompt_hash`), a regex list whose matches are the prompt's dynamic parts. Distinct from `checkpoints/system_prompt.rs`, which resolves a single stripping regex inline on the checkpoints path — this pipeline is span-ingest-wide and cache-first.
+- Flow: the ingest producer (`traces/producer.rs::publish_span_messages`) collects `(prompt_hash, system_prompt)` from `preprocess_for_queue`, checks `static_prompt_regex:{project}:{hash}` and publishes misses to `static_prompt_queue` — best-effort, and only AFTER the span publish succeeded so rejected batches don't feed the accumulator. The consumer (`traces/system_extraction/consumer.rs`) appends the prompt to `static_prompt_accumulator:{project}:{hash}` (24h TTL, capped at 10 samples so a persistently failing agent can't grow the cache value unboundedly); at ≥ 5 samples it takes the per-signature lock `static_prompt_lock:{project}:{hash}` (30-min TTL — runs are normally under 10 min, but the per-step upper bounds are high just in case), runs the extraction agent, writes the regex cache (7-day TTL), clears the accumulator, and releases the lock.
 - On AGENT failure the lock is deliberately NOT released — TTL expiry rate-limits retries against a failing agent. On a post-agent regex-cache WRITE failure the lock IS released (the agent already succeeded; holding the lock would burn a wasted agent call waiting out the TTL). Lock contention and cache-read errors both drop the message (no requeue); a later span with the same signature re-triggers.
 - The consumer calls the real agent, `traces::system_extraction::extract_static_regexes`. It returns `ExtractionResult` (never errors); an EMPTY `regexes` list means every temperature episode failed and the consumer treats it as agent failure (no cache write, lock kept). The agent needs an `LlmClient` — `StaticPromptHandler` carries `Option<Arc<LlmClient>>` (built in main.rs only when `Feature::Signals` is enabled) and drains the queue without accumulating when it's `None`. Tests use the `#[cfg(test)] test_regexes` seam on the handler instead of a real client.
 - The read-modify-write on the accumulator is NOT atomic across workers — concurrent appends to the same signature can lose samples. Acceptable: the threshold is a heuristic, and the stream of spans keeps feeding it.
